refactor(locks): move the content-lock publish workflow into LocksApplication (#2026)
